Rewritten GUI, feature add
Didn't like how the old GUI looked and functioned. I rewrote it to feature a more streamlined experience. It should be less finicky to use now.
Added an info button for nerds that you can click on. It will show the entropy values of the candidate passwords and how the occurrence map looks like. More to satisfy my own curiosity than actual functional use.
I also found a pretty major bug that incorrectly updated the GUI whenever candidate passwords were removed. Basically the suggestive selection of the JSpinner was wrong after reporting the first hitFeedback. This is a bug in my ActionListener implementation and I did NOT fix it since I've rewritten the whole thing anyway and the new GUI doesn't have that problem. Instead, I set the old GUI as deprecated.